final version graph still with some bugs
