+++ app/src/main/java/com/example/mint/controller/ItineraryActivity.java
@@ -23,7 +23,6 @@ import com.example.mint.R;
 import com.example.mint.model.Itinerary;
 import com.example.mint.model.PreferencesAddresses;
 import com.example.mint.model.PreferencesPollution;
-import com.example.mint.model.PreferencesPollution2;
 import com.example.mint.model.PreferencesSensibility;
 import com.example.mint.model.PreferencesSize;
 import com.example.mint.model.Step;
@@ -748,11 +747,11 @@ public class ItineraryActivity extends AppCompatActivity implements View.OnClick
                             "L'exposition associée à ce trajet a bien été ajoutée à votre profil",
                             Toast.LENGTH_SHORT).show();
                     //get the pollution already saved today
-                    int pollution = PreferencesPollution2.getPollutionToday(ItineraryActivity.this);
+                    int pollution = PreferencesPollution.getPollutionToday(ItineraryActivity.this);
                     //add the pollution of this itinerary
                     pollution+=(int) list.get(i).getPollution();
                     // save it in the preferences
-                    PreferencesPollution2.setPollutionToday(pollution, ItineraryActivity.this);
+                    PreferencesPollution.setPollutionToday(pollution, ItineraryActivity.this);
                 }
             });
