+++ app/src/main/java/com/example/mint/model/PreferencesDate.java
@@ -1,80 +1,31 @@
 package com.example.mint.model;
 
-import android.content.Context;
-import android.content.SharedPreferences;
-
+import java.text.SimpleDateFormat;
+import java.time.YearMonth;
+import java.util.ArrayList;
 import java.util.Calendar;
 
+import android.annotation.SuppressLint;
+import android.content.Context;
+import android.content.SharedPreferences;
+import android.os.Build;
+
+import androidx.annotation.RequiresApi;
+
 public class PreferencesDate {
+    @SuppressLint("SimpleDateFormat")
+    static SimpleDateFormat formatDate = new SimpleDateFormat("d_M_yy");
 
     /**
-     * TODO comment
-     *
-     * @param context
+     * Method to get the current date on the format "d_M_yy" (7_2_22) for the 7th of february 2022
+     * @return String date
      */
-    public static void setDate(Context context) {
-        SharedPreferences prefs = context.getSharedPreferences("date", Context.MODE_PRIVATE);
-        SharedPreferences.Editor editor = prefs.edit();
+    public static String getCurrentDate() {
         final Calendar cldr = Calendar.getInstance();
-        int year = cldr.get(Calendar.YEAR);
-        int month = cldr.get(Calendar.MONTH);
-        int day = cldr.get(Calendar.DAY_OF_MONTH);
-        editor.putInt("year", year);
-        editor.putInt("month", (month + 1));
-        editor.putInt("day", day);
-        editor.apply();
-    }
-
-    /**
-     * this returns the last date we stored
-     *
-     * @param context
-     * @return
-     */
-    public static int[] getLastDate(Context context) {
-        // this is going to be used to check whether we have started a new day or month or year since the date was last saved
-        SharedPreferences prefs = context.getSharedPreferences("date", Context.MODE_PRIVATE);
-
-        // we are going to store the date as so : {day,month,year}
-        int[] lastDate = new int[3];
-        lastDate[0] = prefs.getInt("day", 0);
-        lastDate[1] = prefs.getInt("month", 0);
-        lastDate[2] = prefs.getInt("year", 0);
-        return lastDate;
-    }
-
-    /**
-     * TODO comment
-     *
-     * @return
-     */
-    public static int[] getCurrentDate() {
-        // this doesn't use sharedPreferences but we put it here anyway to have all the time handling methods in the same place
-        final Calendar cldr = Calendar.getInstance();
-        int year = cldr.get(Calendar.YEAR);
-        int month = cldr.get(Calendar.MONTH);
-        int day = cldr.get(Calendar.DAY_OF_MONTH);
-        int dayOfWeek = cldr.get(Calendar.DAY_OF_WEEK);
-        int daysInMonth = cldr.getActualMaximum(Calendar.DAY_OF_MONTH);
-        int[] date = {day, (month + 1), year, dayOfWeek, daysInMonth};
+        final String date = formatDate.format(cldr.getTime());
         return date;
     }
 
-    /**
-     * This method checks if the date is the last one in the given month, if so return false, else return true
-     *
-     * @param date the given date
-     * @return
-     */
-    public static boolean checkIfSameMonth(int[] date) {
-        if ((date[0] == 31 && (date[1] == 1 || date[1] == 3 || date[1] == 5 || date[1] == 7 || date[1] == 8 || date[1] == 10 || date[1] == 12))) {
-            return false;
-        } else if ((date[0] == 30 && (date[1] == 4 || date[1] == 6 || date[1] == 9 || date[1] == 11))) {
-            return false;
-        } else if (date[0] == 28 && date[1] == 2) {
-            return false;
-        } else {
-            return true;
-        }
-    }
 }
+
+
